TemplatesHelper moved to Generator\ directory. Let `templates` means just templates.
